rename: 스토어 폴더명 변경
